projects/isitup/deployments/docker/isitup/requirements.txt
certifi==2020.4.5.1
chardet==3.0.4
click==7.1.2
Flask==1.1.2
idna==2.9
itsdangerous==1.1.0
Jinja2==2.11.2
MarkupSafe==1.1.1
PyYAML==5.3.1
requests==2.23.0
urllib3==1.25.9
Werkzeug==1.0.1
PyGithub==1.54.1
mysqlclient==1.4.6